requirements.txt
Django==2.2.10
djangorestframework==3.9.1
django-filter==2.1.0
django-rest-swagger==2.2.0